+++ src/browser.mjs
@@ -1,7 +1,7 @@
 // Core
 //
 export { against, match, when, otherwise, pluck } from './match-iz.mjs'
-export { not, anyOf, allOf, spread } from './match-iz.mjs'
+export { eq, deepEq, not, anyOf, allOf, spread } from './match-iz.mjs'
 export { cata, instanceOf, hasOwn } from './match-iz.mjs'
 export { defined, empty, truthy, falsy } from './match-iz.mjs'
 export { startsWith, endsWith, includes, includedIn } from './match-iz.mjs'